refactor common parts in sql and parquet
