+++ cloud2sql/sql.py
@@ -1,36 +1,33 @@
-import inspect
 from functools import lru_cache
-from typing import List, Any, Type, Tuple, Dict, Optional
+from typing import Any, Type, Dict, Optional
 
-from resotoclient.models import Kind, Model, Property, JsObject
-from resotolib import baseresources
+from resotoclient.models import Kind, Model, JsObject
 from resotolib.args import Namespace
-from resotolib.baseresources import BaseResource
-from resotolib.types import Json
-from sqlalchemy import Boolean, Column, Float, Integer, JSON, MetaData, String, Table, DDL
+from sqlalchemy import (
+    Boolean,
+    Column,
+    Float,
+    Integer,
+    JSON,
+    MetaData,
+    String,
+    Table,
+    DDL,
+)
 from sqlalchemy.engine import Engine, Connection
 from sqlalchemy.sql import Insert
 from sqlalchemy.sql.ddl import DropTable, DropConstraint
 from sqlalchemy.sql.dml import ValuesBase
 from sqlalchemy.sql.type_api import TypeEngine
 
-from cloud2sql.util import value_in_path
-
-# This set will hold the names of all "base" resources
-# Since that are abstract classes, there will be no instances of them - hence we do not need a table for them.
-base_kinds = {
-    clazz.kind for _, clazz in inspect.getmembers(baseresources, inspect.isclass) if issubclass(clazz, BaseResource)
-}
-
-temp_prefix = "tmp_"
-
-carz = [
-    Property("cloud", "string"),
-    Property("account", "string"),
-    Property("region", "string"),
-    Property("zone", "string"),
-]
-carz_access = {name: ["ancestors", name, "reported", "id"] for name in ["cloud", "account", "region", "zone"]}
+from cloud2sql.schema_utils import (
+    base_kinds,
+    get_table_name,
+    get_link_table_name,
+    kind_properties,
+    temp_prefix,
+    insert_node,
+)
 
 
 class SqlModel:
@@ -62,47 +59,13 @@ class SqlModel:
         else:
             return JSON
 
-    def table_name(self, kind: str, with_tmp_prefix: bool = True) -> str:
-        replaced = kind.replace(".", "_")
-        return temp_prefix + replaced if with_tmp_prefix else replaced
-
-    def link_table_name(self, from_kind: str, to_kind: str, with_tmp_prefix: bool = True) -> str:
-        # postgres table names are not allowed to be longer than 63 characters
-        replaced = f"link_{self.table_name(from_kind, False)[0:25]}_{self.table_name(to_kind, False)[0:25]}"
-        return temp_prefix + replaced if with_tmp_prefix else replaced
-
-    def kind_properties(self, kind: Kind) -> Tuple[List[Property], List[str]]:
-        visited = set()
-
-        def base_props_not_visited(kd: Kind) -> Tuple[List[Property], List[str]]:
-            if kd.fqn in visited:
-                return [], []
-            visited.add(kd.fqn)
-            # take all properties that are not synthetic
-            # also ignore the kind property, since it is available in the table name
-            properties: Dict[str, Property] = {
-                prop.name: prop for prop in (kd.properties or []) if prop.synthetic is None and prop.name != "kind"
-            }
-            defaults = kd.successor_kinds.get("default") if kd.successor_kinds else None
-            successors: List[str] = defaults.copy() if defaults else []
-            for kind_name in kd.bases or []:
-                if ck := self.model.kinds.get(kind_name):  # and kind_name != "resource":
-                    props, succs = base_props_not_visited(ck)
-                    for prop in props:
-                        properties[prop.name] = prop
-                    successors.extend(succs)
-            return list(properties.values()), successors
-
-        prs, scs = base_props_not_visited(kind)
-        return prs + carz, scs
-
     def create_schema(self, connection: Connection, args: Namespace) -> MetaData:
         def table_schema(kind: Kind) -> None:
-            table_name = self.table_name(kind.fqn)
+            table_name = get_table_name(kind.fqn)
             if table_name not in self.metadata.tables:
-                properties, _ = self.kind_properties(kind)
+                properties, _ = kind_properties(kind, self.model)
                 Table(
-                    self.table_name(kind.fqn),
+                    get_table_name(kind.fqn),
                     self.metadata,
                     *[
                         Column("_id", String, primary_key=True),
@@ -111,9 +74,9 @@ class SqlModel:
                 )
 
         def link_table_schema(from_kind: str, to_kind: str) -> None:
-            from_table = self.table_name(from_kind)
-            to_table = self.table_name(to_kind)
-            link_table = self.link_table_name(from_kind, to_kind)
+            from_table = get_table_name(from_kind)
+            to_table = get_table_name(to_kind)
+            link_table = get_link_table_name(from_kind, to_kind)
             if (
                 link_table not in self.metadata.tables
                 and from_table in self.metadata.tables
@@ -121,10 +84,15 @@ class SqlModel:
             ):
                 # defining a foreign key constraint on a table that does not exist yet, will fail
                 # defining it on the current tmp table setup is possible, but this will not reset during table rename
-                Table(link_table, self.metadata, Column("from_id", String), Column("to_id", String))
+                Table(
+                    link_table,
+                    self.metadata,
+                    Column("from_id", String),
+                    Column("to_id", String),
+                )
 
         def link_table_schema_from_successors(kind: Kind) -> None:
-            _, successors = self.kind_properties(kind)
+            _, successors = kind_properties(kind, self.model)
             # create link table for all linked entities
             for successor in successors:
                 link_table_schema(kind.fqn, successor)
@@ -180,20 +148,4 @@ class SqlUpdater:
         return maybe_insert.values(values) if maybe_insert is not None else None
 
     def insert_node(self, node: JsObject) -> Optional[ValuesBase]:
-        if node.get("type") == "node" and "id" in node and "reported" in node:
-            reported: Json = node.get("reported", {})
-            reported["_id"] = node["id"]
-            reported["cloud"] = value_in_path(node, carz_access["cloud"])
-            reported["account"] = value_in_path(node, carz_access["account"])
-            reported["region"] = value_in_path(node, carz_access["region"])
-            reported["zone"] = value_in_path(node, carz_access["zone"])
-            kind = reported.pop("kind")
-            self.kind_by_id[node["id"]] = kind
-            return self.insert_value(self.model.table_name(kind), reported)
-        elif node.get("type") == "edge" and "from" in node and "to" in node:
-            from_id = node["from"]
-            to_id = node["to"]
-            if (from_kind := self.kind_by_id.get(from_id)) and (to_kind := self.kind_by_id.get(to_id)):
-                link_table = self.model.link_table_name(from_kind, to_kind)
-                return self.insert_value(link_table, {"from_id": from_id, "to_id": to_id})
-        raise ValueError(f"Unknown node: {node}")
+        return insert_node(node, self.kind_by_id, self.insert_value)
